Backend/requirements.txt
boto3==1.34.67
botocore==1.34.67
Flask==3.0.2
Flask-Cors==4.0.0
mysql-connector-python==8.3.0
